# Fix `getSnippetHash` not considering all files