Made a stupid mistake in the previous PR: https://github.com/grafana/tanka/pull/759

This fixes it and adds another benchmark test to ensure it doesn't happen again. I also removed the Github Actions benchmark test, as it's not really useful, anytime we change the tests, we'll get erroneous results which will be annoying. Instead, I added the benchmark tests to the Drone run, we can compare whenever we want.

Benchmark comparison with v0.22.1:

```
$ benchstat test-0.22.1.txt test-new.txt 
name                                     old time/op    new time/op    delta
pkg:github.com/grafana/tanka/pkg/jsonnet goos:darwin goarch:amd64
GetSnippetHash/all-imported-from-main-8    40.6ms ±10%    30.1ms ± 4%  -25.81%  (p=0.008 n=5+5)
GetSnippetHash/deeply-nested-8             42.7ms ± 4%    30.0ms ± 6%  -29.64%  (p=0.008 n=5+5)

name                                     old alloc/op   new alloc/op   delta
pkg:github.com/grafana/tanka/pkg/jsonnet goos:darwin goarch:amd64
GetSnippetHash/all-imported-from-main-8    15.4MB ± 2%     6.2MB ± 0%  -60.06%  (p=0.008 n=5+5)
GetSnippetHash/deeply-nested-8             15.6MB ± 1%     6.2MB ± 0%  -60.69%  (p=0.008 n=5+5)

name                                     old allocs/op  new allocs/op  delta
pkg:github.com/grafana/tanka/pkg/jsonnet goos:darwin goarch:amd64
GetSnippetHash/all-imported-from-main-8      152k ± 2%       47k ± 0%  -68.86%  (p=0.008 n=5+5)
GetSnippetHash/deeply-nested-8               154k ± 1%       47k ± 0%  -69.17%  (p=0.008 n=5+5)
```

whereas the same test just fails on the `main` branch. I also verified with our largest projects here at Grafana Labs, and the hashes are now consistent with v0.22.1. Runtime goes from 1.8s to 200ms